requirements.txt
# Core ML and Data Science
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
catboost>=1.1.0
joblib>=1.2.0

# LLM Integration
openai>=1.0.0
python-dotenv>=0.19.0

# Web Interface
streamlit>=1.25.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Telegram Bot
python-telegram-bot>=20.0

# Database
sqlite3  # Built-in with Python

# Additional utilities
requests>=2.28.0 